Desigh plugin system for Megatron-LM-FL
update

update

plugins for optimizers

add logger

fix

fix errors

fix

fix plugins functions call

fix bugs

fix import bugs

add mixed optimizer plugins

remove md

fix

test dist ckpt

remove print

fix typo error

update

update hetero

update

update language module

update plugins

update tokenizer

add logging

fix import

rename plugins to plugin

fix

fix import

polish

fix

rn

fix cur import

fix

add lazy load

fix

fix

fix

fix
