+++ megatron/core/datasets/blended_dataset.py
@@ -13,8 +13,10 @@ import torch
 
 from megatron.core.datasets.blended_megatron_dataset_config import BlendedMegatronDatasetConfig
 from megatron.core.datasets.megatron_dataset import MegatronDataset
-from megatron.core.datasets.utils import normalize, is_built_on_zero_rank
+from megatron.core.datasets.utils import normalize
 from megatron.core.utils import log_single_rank
+
+from plugin.core.datasets.utils import is_built_on_zero_rank
 
 logger = logging.getLogger(__name__)
