+++ megatron/core/transformer/moe/moe_utils.py
@@ -12,6 +12,8 @@ from megatron.core.process_groups_config import ProcessGroupCollection
 from megatron.core.transformer.cuda_graphs import is_graph_capturing
 from megatron.core.transformer.transformer_config import TransformerConfig
 
+from plugin.core.transformer.moe.moe_utils import reduce_aux_losses_tracker_across_ranks_hetero
+
 try:
     import transformer_engine as te  # pylint: disable=unused-import
 
@@ -786,37 +788,7 @@ def reduce_aux_losses_tracker_across_ranks(track_names: Optional[List[str]] = No
                 values, group=tracker[name]['avg_group'], op=torch.distributed.ReduceOp.AVG
             )
 
-def reduce_aux_losses_tracker_across_ranks_hetero(
-    track_names: Optional[List[str]] = None,
-):
-    """Collect and reduce the auxiliary losses across ranks."""
-    tracker = parallel_state.get_moe_layer_wise_logging_tracker()
-    if track_names is None:
-        track_names = tracker.keys()
-    for name in track_names:
-        values = tracker[name]["values"]
-        # Reduce aux losses across ranks.
-        if tracker[name].get("reduce_group") is not None:
-            torch.distributed.all_reduce(
-                values, group=tracker[name].get("reduce_group")
-            )
-        if tracker[name].get("avg_group") is not None:
-            torch.distributed.all_reduce(
-                values,
-                group=tracker[name]["avg_group"],
-                op=torch.distributed.ReduceOp.AVG,
-            )
-        pp_groups = parallel_state.get_pipeline_model_parallel_group()
-        if "cpu:gloo" == torch.distributed.get_backend(pp_groups[0]):
-            values = values.cpu()
-        assert isinstance(pp_groups, list), "pp_groups should be a list for hetero."
-        if len(pp_groups) > 1:
-            origin_values = values.clone().detach()
-            for pp_group in pp_groups:
-                values.copy_(origin_values)
-                torch.distributed.all_reduce(values, group=pp_group)
-        else:
-            torch.distributed.all_reduce(values, group=pp_groups[0])
+
 
 def track_moe_metrics(
     loss_scale: float,
