+++ megatron/training/global_vars.py
@@ -14,7 +14,7 @@ from megatron.core.num_microbatches_calculator import init_num_microbatches_calc
 from megatron.training.dist_signal_handler import DistributedSignalHandler
 from megatron.training.tokenizer import build_tokenizer
 
-from flagscale.train.hetero.p2p_communication import get_device_type_for_comm
+from plugin.hetero.p2p_communication import get_device_type_for_comm
 
 _GLOBAL_ARGS = None
 _GLOBAL_TOKENIZER = None
