+++ megatron/training/utils.py
@@ -43,8 +43,7 @@ from megatron.core.utils import (
 )
 from megatron.legacy.model.module import param_is_not_shared
 
-from flagscale.train.hetero.p2p_communication import get_device_type_for_comm # FlagScale Add
-
+from plugin.hetero.p2p_communication import get_device_type_for_comm
 
 def calc_params_l2_norm(model, force_create_fp32_copy=False):
     """Calculate l2 norm of parameters"""
