+++ plugin/CIRCULAR_IMPORT_ANALYSIS.md
@@ -1,0 +1,114 @@
+# 延迟导入机制循环导入分析
+
+## 当前设计
+
+### 延迟导入机制
+
+```python
+# plugin/decorators.py
+@plugin_method
+def wrapper(*args, **kwargs):
+    plugin_impl = get_plugin_method(method_key)
+    
+    # 如果找不到，延迟导入 plugin 模块
+    if plugin_impl is None:
+        plugin_module = megatron_module.replace("megatron.", "plugin.", 1)
+        importlib.import_module(plugin_module)  # 延迟导入
+        plugin_impl = get_plugin_method(method_key)
+```
+
+### 执行时机
+
+1. **模块导入时**：
+   - megatron 模块初始化
+   - `@plugin_method` 装饰器执行（只是装饰，不执行函数体）
+   - **不导入 plugin 模块** ✅
+
+2. **函数调用时**（运行时）：
+   - megatron 模块已经完全初始化 ✅
+   - `wrapper` 函数执行
+   - 如果找不到 plugin，才导入 plugin 模块
+   - 此时导入 `plugin.core.distributed.finalize_model_grads`
+
+## 循环导入分析
+
+### 场景 1: 正常情况（无循环导入）
+
+```
+T1: megatron.core.distributed.finalize_model_grads 初始化
+    → 定义 @plugin_method 装饰的函数
+    → 不导入 plugin ✅
+
+T2: 应用运行，调用 _allreduce_embedding_grad()
+    → wrapper 函数执行
+    → 找不到 plugin，延迟导入 plugin.core.distributed.finalize_model_grads
+    → 此时 megatron 模块已完全初始化 ✅
+    → plugin 文件导入 megatron 模块（第 13-21 行）
+    → 可以正常导入 ✅
+```
+
+**结果：无循环导入** ✅
+
+### 场景 2: 潜在问题（如果 plugin 在模块级别导入）
+
+如果 `plugin.core.distributed.finalize_model_grads` 在模块级别导入：
+```python
+from megatron.core.distributed.finalize_model_grads import _get_main_grad_attr
+```
+
+**执行流程**：
+```
+T1: megatron.core.distributed.finalize_model_grads 初始化中
+    → 定义 @plugin_method 装饰的函数
+    → 不导入 plugin ✅
+
+T2: 应用运行，调用 _allreduce_embedding_grad()
+    → wrapper 函数执行
+    → 延迟导入 plugin.core.distributed.finalize_model_grads
+    → plugin 文件第 21 行：from megatron.core.distributed.finalize_model_grads import ...
+    → 此时 megatron.core.distributed.finalize_model_grads 已完全初始化 ✅
+    → 可以正常导入 ✅
+```
+
+**结果：无循环导入** ✅
+
+## 为什么不会有循环导入？
+
+### 关键点
+
+1. **延迟导入在函数调用时执行**，不是在模块导入时
+2. **函数调用时，megatron 模块已完全初始化**
+3. **plugin 文件导入 megatron 时，megatron 已经完成初始化**
+
+### 时间线
+
+```
+模块导入阶段：
+  megatron.core.distributed.finalize_model_grads 初始化
+    → 定义函数，装饰器执行（只装饰，不执行函数体）
+    → 不导入 plugin ✅
+  
+运行时（函数调用）：
+  调用 _allreduce_embedding_grad()
+    → wrapper 执行
+    → 延迟导入 plugin.core.distributed.finalize_model_grads
+    → 此时 megatron 已完全初始化 ✅
+    → plugin 导入 megatron，无问题 ✅
+```
+
+## 结论
+
+**不会有循环导入问题**，因为：
+
+1. ✅ 延迟导入在**函数调用时**执行，不是在模块导入时
+2. ✅ 函数调用时，megatron 模块已完全初始化
+3. ✅ plugin 文件导入 megatron 时，megatron 已经完成初始化
+4. ✅ 即使 plugin 文件在模块级别直接导入 megatron，也不会有问题
+
+## 注意事项
+
+虽然不会有循环导入，但建议：
+
+1. **plugin 文件中的 megatron 导入**：可以在模块级别直接导入（因为延迟导入时 megatron 已初始化）
+2. **但为了更安全**：如果担心，可以在函数内部导入，或者使用 lazy_import
+
