--- plugins/core/__init__.py
@@ -1,8 +1,0 @@
-"""
-Plugin implementations for megatron.core modules.
-"""
-
-# Import plugin implementations
-# The implementations will be registered automatically when the module is imported
-from . import models  # noqa: F401
-
