--- plugins/core/models/__init__.py
@@ -1,6 +1,0 @@
-"""
-Plugin implementations for megatron.core.models modules.
-"""
-
-from . import common  # noqa: F401
-
