--- plugins/core/models/common/__init__.py
@@ -1,6 +1,0 @@
-"""
-Plugin implementations for megatron.core.models.common modules.
-"""
-
-from . import language_module  # noqa: F401
-
