--- plugins/core/models/common/language_module/__init__.py
@@ -1,10 +1,0 @@
-"""
-Plugin implementations for megatron.core.models.common.language_module.
-
-This module automatically registers plugin implementations when imported.
-"""
-
-# Import the plugin implementation
-# The decorator will automatically register when the module is imported
-from . import language_module  # noqa: F401
-
